refactor(FE): Toastr component
