+++ frontend/src/components/Toastr/index.js
@@ -1,0 +1,2 @@
+export * from './Toastr';
+export { default } from './Toastr';
